requirements.txt
bangla==0.0.2
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.2.3
django-auditlog==0.4.5
django-cors-headers==3.0.2
django-environ==0.4.5
django-jinja==2.4.1
django-jsonfield==1.3.1
django-oauth-toolkit==1.2.0
django-rest-passwordreset==1.0.0
django-shortuuidfield==0.1.3
django-smalluuid==1.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.2
openpyxl==2.6.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.6.0
pytz==2019.1
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
smalluuid==1.0.0
sqlparse==0.3.0
tinycss2==1.0.2
urllib3==1.25.3
WeasyPrint==48
webencodings==0.5.1
